Removed Interface 'DataProvider', removed File_DataProvider, use DB Mode only. Implemented some SQL-Select statements in DB_DataProvider

+++ Costimizer/shopitem.cpp
@@ -1,17 +1,17 @@
 #include "shopitem.h"
 
 ShopItem::ShopItem()
-: ShopItem{ 0, "" }
+: ShopItem{ 0u, "" }
 {
 }
 
-ShopItem::ShopItem( const ulong &id, const QString &name )
+ShopItem::ShopItem( const uint &id, const QString &name )
 : id{ id }
 , name{ name }
 {
 }
 
-const ulong& ShopItem::getId() const
+const uint& ShopItem::getId() const
 {
     return this->id;
 }
